NAT and IVA fragment part is completed

+++ app/src/main/java/com/example/quizapp/backend/Question/NAT.java
@@ -56,7 +56,7 @@ public class NAT extends Question implements Serializable {
         return question;
     }
 
-    public double getAnswer() {
+    public float getAnswer() {
         return answer;
     }
